It is propose a more general ConvRNN Cells that includes ConvJANET Cells, also too Chrono initializer is added by default.
In order to do a study for the thesis "Convolutional Recurrent Nervous Networks for Remaining Useful Life Prediction in Mechanical Systems", the code here is used. Among the particularities of this code is that the Conv1DLSTMCell, Conv2DLSTMCell and Conv3DLSTMCell classes are eliminated, for convenience and because they are redundant with the conv_ndims option. In addition, the Chrono initializer is added to the forget gate of the ConvLSTM and ConvJANET, and the Xavier initializer is also placed by default for the rest.

In the study it is concluded that the ConvJANET cell obtains similar results as a ConvLSTM (a little better in certain cases) for the case of prediction from operational data of machines, needing much less computational capacity.